friendly view added for last_seen

+++ social-otter/app/models/social.py
@@ -9,4 +9,4 @@ class Tweet(BaseModel):
     username: str
     displayname: str
     profileImageUrl: str
-    tweet_at: int
+    tweet_at: float
